Restructured anonymize.py
